backend/requirements.txt
Django==4.2.15
djangorestframework==3.14.0
drf-spectacular==0.27.2
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
python-decouple==3.8
celery==5.3.1
redis==4.6.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
transformers==4.33.2
torch==2.0.1
nltk==3.8.1
spacy==3.6.1
# Testing and development
pytest==7.4.0
pytest-django==4.5.2
coverage==7.3.0
gunicorn==21.2.0

# Development tools
django-extensions==3.2.3
watchdog==3.0.0